refactor: GatheringListItem -> GatheringPreview  - 모임 목록 DTO 이름 변경 및 멤버 리스트 필드 추가  - querydsl + join이 필요하므로 DTO 클래스 Core 모듈로 이동
